moved Badge to deprecated folder, and created fake Badge pointer for backward compatibility
