+++ packages/wix-ui-core/src/components/deprecated/Badge/Badge.e2e.ts
@@ -1,7 +1,7 @@
 import * as eyes from 'eyes.it';
 import {browser} from 'protractor';
 import {getStoryUrl, waitForVisibilityOf} from 'wix-ui-test-utils/protractor';
-import {badgeTestkitFactory} from '../../testkit/protractor';
+import {badgeTestkitFactory} from '../../../testkit/protractor';
 
 // Badge component is to be removed and no longer has automated story,
 // thus this test fails.
