+++ packages/wix-ui-core/src/components/deprecated/Badge/styles.ts
