+++ packages/wix-ui-core/src/components/deprecated/Badge/theme.ts
@@ -1,4 +1,4 @@
-import {palette} from '../../palette';
+import {palette} from '../../../palette';
 
 export type BadgeState = Partial<{
   color: string;
