+++ packages/wix-ui-core/src/testkit/index.ts
@@ -11,7 +11,7 @@ export const paginationTestkitFactory = testkitFactoryCreator(
   paginationDriverFactory
 );
 
-import { badgeDriverFactory } from "../components/Badge/Badge.driver";
+import { badgeDriverFactory } from "../components/deprecated/Badge/Badge.driver";
 export const badgeTestkitFactory = testkitFactoryCreator(badgeDriverFactory);
 
 import { tooltipDriverFactory } from "../components/Tooltip/Tooltip.driver";
